Fix issue with mypy checking simulators/mujoco by renaming simulators/mujoco to simulators/mujoco_simulator.
Mypy was confusing mujoco and our package. Renaming was the easiest
solution.
