+++ codetools/mypy/mypy.ini
@@ -11,8 +11,7 @@ ignore_missing_imports = True
 ignore_missing_imports = True
 
 [mypy-mujoco.*]
-# Mypy gets confused by the wildcard imports in the top level mujoco file.
-ignore_errors = True
+ignore_missing_imports = True
 
 [mypy-mujoco_viewer.*]
 ignore_missing_imports = True
