+++ codetools/packages.txt
@@ -16,4 +16,4 @@ experimentation/revolve2
 modular_robot/revolve2
 modular_robot_simulation/revolve2
 simulation/revolve2
-simulators/mujoco/revolve2
+simulators/mujoco_simulator/revolve2
