+++ examples/custom_brain/main.py
@@ -12,12 +12,12 @@ from revolve2.modular_robot.body import RightAngles
 from revolve2.modular_robot.body.v1 import ActiveHingeV1, BodyV1, BrickV1
 from revolve2.modular_robot.brain import Brain, BrainInstance
 from revolve2.modular_robot_simulation import ModularRobotScene, simulate_scenes
-from revolve2.simulators.mujoco import LocalSimulator
+from revolve2.simulators.mujoco_simulator import LocalSimulator
 
 
-def make_body() -> tuple[
-    BodyV1, ActiveHingeV1, ActiveHingeV1, ActiveHingeV1, ActiveHingeV1
-]:
+def make_body() -> (
+    tuple[BodyV1, ActiveHingeV1, ActiveHingeV1, ActiveHingeV1, ActiveHingeV1]
+):
     """
     Create a body for the robot.
