+++ examples/evaluate_multiple_interacting_robots/main.py
@@ -10,7 +10,7 @@ from revolve2.modular_robot import ModularRobot
 from revolve2.modular_robot.brain.cpg import BrainCpgNetworkNeighborRandom
 from revolve2.modular_robot_simulation import ModularRobotScene, simulate_scenes
 from revolve2.simulation.scene import Pose
-from revolve2.simulators.mujoco import LocalSimulator
+from revolve2.simulators.mujoco_simulator import LocalSimulator
 
 
 def main() -> None:
