+++ simulators/mujoco_simulator/revolve2/simulators/mujoco_simulator/_scene_to_model.py
