+++ simulators/mujoco_simulator/revolve2/simulators/mujoco_simulator/_simulate_scene.py
